[closes #35] Added option to have multiple storage paths
This requires to create a new config file. The old config file will be loaded as before, but all path settings will be ignore and required to be added again. Each path now has a ignore list attached to it. Users can also choose which paths to search in and what ignorelists to use.